README.md
A referral code, also known as a promo code, signup bonus code, or invitation code, is a special string you enter during registration on BingX to unlock exclusive trading rewards. In 2025, the BingX referral code IMOIJ1 gives new users a welcome bonus of up to 5125 USDT along with a 50% discount on trading fees.